# openapi/admin.yaml
openapi: 3.0.3
info:
  title: one-llm-router Admin API
  version: 0.5.0
  description: |
    Admin control plane — Settings, OAuth flow, accounts, auth.json
    import/export, request history, Account Playground, and dashboard
    observability.
    Authoritative source of truth for the Settings surface migrated
    from **002-setup-wizard-and-admin-portal-skeleton** and the admin
    surfaces added by **003-multi-mode-codex-auth** and later
    features, per `AGENTS.md §API Contract` rule #1.

    All `/api/admin/*` operations in this document follow the **HTTP 200
    + envelope** policy defined in `docs/error-codes.md` §HTTP envelope
    policy:

      * Both success AND business errors respond with HTTP 200. The
        outcome lives in `body.code` (0 = success, non-zero = business
        error).
      * HTTP 500 is reserved for system errors (`oauth_internal_error` /
        `oauth_store_failed` / `db_unavailable`).
      * HTTP 4xx is FORBIDDEN on any admin route. The envelope moves
        business-error signalling into `body.code`.

    **Documented exemptions (NOT in this file)**:

      1. `GET /auth/callback` — the browser-facing loopback receiver of
         Rail A (OAuth browser flow). Not under `/api/*` and not served
         by the admin mux; prose lives in
         `specs/003-multi-mode-codex-auth/contracts/oauth-flow-api.md`.
      2. `POST /api/admin/accounts/{id}/export-auth-json` **success**
         body — raw Codex CLI `auth.json` bytes with
...

...

...

...

...

...

...
servers:
...
security:
  - adminAuth: []
# Enforcement status of the global `security` block. Read this before
# wiring generated clients:
#
#   - MVP (001 + 002 + 003):   NOT enforced. The router accepts every
#     /api/admin/* request regardless of whether X-Admin-Token is
#     present. The declaration above is aspirational — it pins the
#     auth shape so client generation is stable when the gate lands.
#   - Phase 2+ (future):       ENFORCED via the admin_auth plugin.
#
# The `x-mvp-enforcement` extension below is read by downstream
# tooling (docs site, client-codegen post-processors) to surface the
# "declared but not yet enforced" disclaimer. AGENTS.md §Security is
# the source of truth for the phased rollout — see "Phase 1 / Phase
# 2+" split in that doc.
...
paths:
  /api/admin/settings:
    get:
      operationId: settingsGet
      tags: [settings]
      summary: Return hot-reloadable runtime settings.
      description: |
        Returns the live Settings page projection: the five
        hot-reloadable runtime controls, non-secret DB identity, live
        plugin summaries, persisted plugin intents, and build metadata.
        `db.url` and every credential-bearing field are intentionally
        omitted.
      responses:
        '200':
          description: Envelope with Settings payload or setup-required business error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsGetResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/settings/update:
    post:
      operationId: settingsUpdate
      tags: [settings]
      summary: Patch hot-reloadable runtime settings and plugin intents.
      description: |
        Partial, RPC-style settings mutation. Only the five
        `runtime.*` fields and the two persisted plugin intent flags
        (`plugins.admin_auth.enabled`, `plugins.client_keys.enabled`)
        are patchable. Unknown keys are rejected with
        `2012 unknown_config_key`. On success the response returns the
        same `data` shape as `GET /api/admin/settings` so the SPA can
        hydrate its cache without a follow-up GET.

        The endpoint has a 16 KiB JSON body cap and reports body-cap
        failures as HTTP 200 + `2009 request_body_too_large` with
        `data.scope="envelope"` and `data.limit_bytes=16384`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettingsPatchRequest' }
      responses:
        '200':
          description: Envelope — success or Settings business error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsUpdateResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/oauth/browser/start:
    post:
      operationId: oauthBrowserStart
      tags: [oauth-flow]
      summary: Start a browser PKCE OAuth flow (dual-rail).
      description: |
        Opens a single browser-OAuth flow per FR-012. The server always
        tries to bind the loopback listener (Rail A) AND always accepts
        a manually-pasted callback URL (Rail B); the UI unconditionally
        renders the paste textarea. Whichever rail delivers a valid
        `code + state` first wins via a `Flow.Consumed` atomic CAS.

        There is no `mode` / `topology` parameter. Unknown JSON keys in
        the request body are silently ignored (forward-compat).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OAuthStartRequest' }
      responses:
        '200':
          description: Envelope with success or business-error payload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrowserStartResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/oauth/browser/manual-callback:
    post:
      operationId: oauthBrowserManualCallback
      tags: [oauth-flow]
      summary: Rail B — submit a manually-pasted callback URL.
      description: |
        Processing order: URL prefix validation → required-param check
        (`code` OR `error` present) → constant-time `state` comparison
        → `Flow.Consumed` CAS → `/oauth/token` code-exchange. **State
        validation runs BEFORE any branch including `error=access_denied`
        — an unauthenticated cancel MUST NOT tear down a live flow.**
        See `contracts/oauth-flow-api.md` §`POST /browser/manual-callback`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualCallbackRequest' }
      responses:
        '200':
          description: Envelope — success, cancel, or per-rail rejection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManualCallbackResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/oauth/device/start:
    post:
      operationId: oauthDeviceStart
      tags: [oauth-flow]
      summary: Start a device-code OAuth flow.
      description: |
        Kicks off a background polling goroutine that hits OpenAI's
        device-auth token endpoint per `research.md` Decision 1
        (endpoint bag + RFC 8628 §3.5 interval fallback). On
        `slow_down` the poller **doubles** the current interval
        (capped at 60s) — NOT "+5s".
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OAuthStartRequest' }
      responses:
        '200':
          description: Envelope — success or business-error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceStartResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/oauth/flow:
    get:
      operationId: oauthFlowStatus
      tags: [oauth-flow]
      summary: Status poll — the only endpoint the UI hits during a flow.
      description: |
        Returns `idle` / `pending` / `success` / `error` inside
        `data.status`. Success and error are **one-shot** — the server
        flips back to idle on the NEXT read so the UI never sees the
        same terminal outcome twice.
      responses:
        '200':
          description: Envelope — always `code:0`; detail lives in `data.status`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FlowStatusEnvelope' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/oauth/cancel:
    post:
      operationId: oauthCancel
      tags: [oauth-flow]
      summary: Cancel the pending flow (operator-initiated teardown).
      description: Idempotent — cancelling an idle flow returns `status:"idle"` with `code:0`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelFlowRequest' }
      responses:
        '200':
          description: Envelope — idle success or `flow_id_mismatch`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthCancelResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
...
  /api/admin/accounts/{id}/export-auth-json:
    post:
      operationId: accountsExportAuthJSON
      tags: [accounts]
      summary: Export an OAuth account row as a Codex CLI `auth.json` attachment.
      description: |
        **Dual-shape success**: the success body is raw Codex CLI
        `auth.json` bytes (byte-compatible with the shape Codex reads)
        served with `Content-Disposition: attachment; filename="auth.json"`
        — the second documented envelope exemption in 003. ERROR
        responses still use the envelope.

        Every hit is audited at WARN level via `oauth_auth_json_exported`
        (success) or `oauth_auth_json_export_rejected` (business
        error). No rate limit (explicit clarification per spec.md
        §Clarifications 2026-04-15).
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '200':
          description: |
            Raw `auth.json` attachment on success, envelope on business
            error. Callers MUST branch on `Content-Type`
            (`application/json` + `Content-Disposition: attachment`
            ⇒ raw; `application/json` without `Content-Disposition`
            ⇒ envelope).
          headers:
            Content-Disposition:
              description: Present on success only — `attachment; filename="auth.json"`.
              schema: { type: string }
            Cache-Control:
              schema: { type: string, example: "no-store, private" }
            X-Content-Type-Options:
              schema: { type: string, example: "nosniff" }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportAuthJSONResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/dashboard:
    get:
      operationId: dashboardGet
      tags: [dashboard]
      summary: Return the Admin Dashboard observability snapshot.
      description: |
        Returns read-only dashboard card metrics derived from retained
        request records and current account state. The default range is
        the latest seven days. Time-ranged cards can be filtered to one
        upstream account; the active-account card always reflects current
        router capacity across all accounts.
      parameters:
        - name: range
          in: query
          schema: { $ref: '#/components/schemas/DashboardRange' }
          description: Time window for time-ranged cards. Defaults to `7d`.
        - name: account_id
          in: query
          schema: { type: integer, format: int64, minimum: 1 }
          description: Optional upstream account id filter for time-ranged cards.
      responses:
        '200':
          description: Envelope with dashboard card data or a filter validation error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardResponseBody' }
        '500':
          $ref: '#/components/responses/SystemError'
  /api/admin/usage:
    get:
      operationId: usageGet
      tags: [usage]
      summary: Return router-local usage summary.
      description: |
        Returns usage counters derived from retained router request
        records, plus a Codex-account status projection derived from
        active OAuth accounts. This is a router-owned Admin API endpoint
        and therefore uses the standard `{code,msg,data}` envelope.
        It is the only router-local usage surface in Feature 006.
      responses:
        '200':
          description: Envelope with router-local usage data.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageEnvelope' }
        '500':
          $ref: '#/components/responses/SystemError'
...
components:
  securitySchemes:
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: |
        Admin-auth plugin (target future feature). 003+ endpoints are
        **ready** to be gated by this scheme but the plugin
        implementation ships with a later feature; until it lands the
        scheme is advisory (no enforcement) so that single-operator
        local deploys continue to work out-of-the-box. See the
        top-level `x-mvp-enforcement: not-enforced` extension for the
        machine-readable disclaimer and tracked-by pointer.
...